+++ config/config.exs
@@ -10,7 +10,7 @@ use Mix.Config
 config :glific,
   ecto_repos: [Glific.Repo],
   provider: Glific.Providers.Gupshup.Message,
-  #provider_worker: Glific.Providers.Gupshup.Worker,
+  # provider_worker: Glific.Providers.Gupshup.Worker,
   provider_worker: Glific.Providers.Glifproxy.Worker,
   provider_id: "gupshup-provider-23",
   provider_limit: 10
